requirements.txt
asgiref==3.7.2
Django==4.2.1
django-adminlte2==0.4.3
django-grappelli==3.0.8
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
pillow==10.2.0
psycopg2-binary==2.9.9
python-slugify==8.0.2
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.3
